[fix] rating 자료형 변경 (Float->Integer)

+++ server-api/src/main/java/com/depromeet/domains/store/dto/response/StoreReviewResponse.java
@@ -11,13 +11,13 @@ public class StoreReviewResponse {
 
     private Long userId;
     private String nickName;
-    private Float rating;
+    private Integer rating;
     private String imageUrl;
     private Integer visitTimes;
     private LocalDateTime visitedAt;
     private String description;
 
-    public static StoreReviewResponse of(Long userId, String nickName, Float rating, String imageUrl, Integer visitTimes, LocalDateTime visitedAt, String description) {
+    public static StoreReviewResponse of(Long userId, String nickName, Integer rating, String imageUrl, Integer visitTimes, LocalDateTime visitedAt, String description) {
         return StoreReviewResponse.builder()
                 .userId(userId)
                 .nickName(nickName)
